pyproject.toml
[project]
name = "bluesky_search"
version = "0.1.0"
requires-python = ">= 3.11.9"
dependencies = [
    "atproto>=0.0.65",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "supabase>=2.0.0",
]
